Change asserts to checks in duk_def_prop_xxx()
Some limitations of duk_hobject_define_property_internal() (which is
ultimately called by duk_def_prop_xxx()) are not obvious and were only
asserted for.  Change the asserts into actual checks so that if there
is a call site which makes an incorrect assumption (perhaps in some
corner case), an error is thrown instead of something unsafe happening.