Avoid double evaluation of compile/eval macro args
String variants referred to both 'src' and 'DUK_STRLEN(src)' which would
cause double evaluation of macro arguments.  For instance, if 'src' was
obtained from a call to a function, the function would be called twice.

Fixed by taking the strlen() inside duk_api_compile.c.  Not the most elegant
thing as it would be nicer to push this to the caller's footprint, but there
is no reasonably way of doing that with C expression syntax.