Add duk_(u)int_fast_t and duk_small_(u)int_fast_t
These are types intended to be used in true fast path code, i.e. the executor.
They are the same as plain duk_(u)int_t and duk_small_(u)int_t but are
guaranteed to be of the C99 "fast" kind -- not sure if this is useful in
practice.