Format string cleanups, allow NULL+%s/%p in debug
Lots of minor string format cleanups (casts mainly).

Add a special check to detect NULL pointers with %s/%p formats in debug
logging to avoid calling the platform sprintf() with a NULL value.  This
is especially important for %s + NULL, which is not portable and may crash
on some platforms.  Debug logging will now format these specially as "NULL"
and avoid calling the platform for formatting, so debug log call sites don't
need to be careful with NULL pointers any more.