Use 'int' for 'duk_int_t' when large enough
The 'int' type is very likely to be a type convenient for the CPU and
compilers are bound to handle it well.  Also change codepoint type to
use duk_int_t instead of a duk_int_fast32_t.

On x64, 'int' is 4 bytes but duk_int_fast32_t is 8 bytes, which has a
somewhat visible impact on at least code footprint, so 'int' is most
likely preferable for x64.