Treat all \xFF strings as internal (also user)
This allows user code to create internal properties too, so that the
internal properties are not returned by e.g. getOwnPropertyNames().

Prior to this change, a user string beginning with \xFF would not be
treated as internal - unless it happened to be a string already used
by Duktape for its own purposes.  This is quite confusing.