Minor fix for 'caller' special behavior
When a base object has the special [[Get]] behavior described in E5.1
Section 15.3.5.4 or 10.6 for the "caller" property, there is a question
whether a TypeError is thrown when the "caller" property value is a
bound strict function (or only when the value is a non-bound strict
function).

This commit changes the behavior so that a TypeError is also thrown
when the "caller" value is a bound function, by ensuring that strict
bound functions have the "strict" internal flag set.