Minor tweak to Object.defineProperties()
Push the internal Object.defineProperty() function only once outside the
loop to work better with large Object.defineProperty() calls.  The
different is quite small in practice.

It would be better to avoid creating a function object in the first place,
but since we call Object.defineProperty() internally we need the original,
unchanged Object.defineProperty() in case the user has overwritten it.