massif-XXX targets for heap use
Print out the massif graph and a grepped "maximum usage" to get some
quick idea of how much memory a certain ecmascript testcase uses at
the moment.  Used for e.g. a hello world memory usage test.  Usage is:
"make massif-test-dev-hello-world" or "make massif-XXX" for any other
test in ecmascript-testcases/.  A few convenience targets like
massif-helloworld are included.

Set --peak-inaccuracy=0.0 for massif run to get as accurate a peak
estimate as possible.  (This is not that important because the measure
itself is not 100% relevant.)

Also adds a testcase needed by massif-helloworld target.