Drop release notes from dist package
RELEASES.rst is only updated in master branch to make it easier to make
and merge fixes to multiple branches.  Including it in the dist package
would mean that it would need to be updated in maintenance branches too
or it would be out-of-date in patch releases.

So, drop release notes from the dist package, and just link to it in the
dist README.