Better C++ static decl fix
When doing a single file build, drop DUK_INTERNAL_DECL declarations which
would otherwise be mapped to "static".  This avoids the problem with C++
where a static data symbol is both forward declared and defined.

With this change static definitions must appear before their first use in
the single file build.  This requires a change in combine_src.py which is
done in a separate commit.