Handpick files to satisfy static data dependencies
Move a few handpicked files from the alphabetical file list to the top of
the list to ensure that static data definitions appear before their first
use in a single file build.

This is necessary because there are no forward declarations for static data
symbols anymore (which was causing C++ issues).  See GH-63.