Rework propdesc conversion to fix a realloc issue
When Object.defineProperties() was used and set/get value was a lightfunc,
duk_to_object() was used for coercion.  After the coercion, a previously
looked up duk_tval pointer was still being used; if the property table of
the property descriptor had been reallocated (pretty difficult to achieve
but possible, e.g. a finalizer triggered by GC could add an entry to it)
the duk_tval pointer could have been stale.

Rework the entire sequence a little bit and fix the issue in the process.