Fix 'volatile' placement for pointers
This fixes an OSX _setjmp() related crash reported by Andreas Oman.

"volatile int *x" is a non-volatile pointer to a volatile int value.
"int * volatile x" is a volatile pointer to a non-volatile int value.
For pointers that may get reassigned, the latter is correct.

Also add a few asserts.