Fix breakpoint trigger condition
The breakpoint trigger condition included a check that if prev_line was zero,
no line transition would ever be detected.  This means that if a breakpoint
is set on the first line of a function, it would never get triggered - not
even after execution had proceeded beyond it.

Change the condition so that prev_line can be zero, and a transition 0->1 is
generated.  This allows breakpoints at line 1 to match.