Separate (marker) unused tag
Reorder tags to accommodate a separate 'unused' tag so that 'undefined' can
become a single tag write (instead of tag + value like booleans).  This is
good because 'undefined' values are involved in e.g. value stack resizes and
are performance relevant.

Also reorder tags so that "is heap allocated" check can be a single bit test
instead of a comparison when using non-packed duk_tval.  This makes every
DECREF potentially faster because an "is heap allocated" test appears in
every DECREF.

Because "unused" is not intended to appear anywhere in actual use (e.g. as
a value stack value, as a property value, etc), "unused" values will fall
into the default clause of DUK_TAG_xxx switch case statements.  Add an assert
to every such default clause that the value is not intended to be "unused".

Remove duk_push_unused() as it should no longer be used.  It was only used
by the debugger protocol; refuse an inbound "unused" value in the debugger.
This is not breaking compatibility because there was no legitimate usage for
the debug client sending requests with "unused" values.