Accept plain buffer for typedarray, buf zeroing
Accept a plain buffer argument in typedarray constructor.  Behavior will be
the same as for a Duktape.Buffer, i.e. the argument is treated as an array-like
initializer.

Zero buffer data explicitly for ArrayBuffer and typedarrays created, even when
DUK_USE_ZERO_BUFFER_DATA option is not set.  Buffer zeroing is explicitly
required by Khronos/ES6.

Node.js Buffer does *not* zero buffer data by default, so this behavior
is not changed here: Node.js Buffers will be zeroed when
DUK_USE_ZERO_BUFFER_DATA is set, and left as garbage when not set.