Add 0th capture to RegExp.prototype bytecode
The bytecode was missing DUK_REOP_SAVE for indices 0 and 1, i.e. the
zeroth capture, which caused the RegExp.prototype to match an empty
input but without providing the zeroth capture which is assumed in
other parts of Duktape internals.  This caused the GH-492 assert failure
and, with asserts disabled, a segfault.