Change Function .toString() format
Emscripten depends on a Function's .toString() having a certain format
(beyond what is guaranteed by the E5 specification).  In particular, it
won't parse Duktape's current format where the function body is replaced
with a comment indicating the function type, e.g.:

    function foo() {/* ecmascript */}

Change this format to indicate the function type using a string literal
(similarly to how directive prologue works):

    function foo() {"ecmascript"}

This is equally informational and will work directly with Emscripten.

Also change handling for anonymous functions: substituting "anon" for an
anonymous function in an Ecmascript function's .toString() output is a bit
misleading.  It also causes an Emscripten issue when a source regexp expects
there to be no name for an anonymous function in the .toString() output.