Fix two debugger detach handling bugs
Fix two separate bugs:

* Potential to re-enter the debugger recursively when a detach happened
  inside the debugger message loop.

* Potential for detach initiated from outside the debugger message loop to
  be never finalized.

The re-entry bug was caused by `dbg_processing` being set to 0 in detach1()
while in the debugger message loop.  This caused the potential for the
debugger code to be re-entered recursively before the detach was finalized
using detach2(), which could have various harmful side effects.  The fix is
to keep `dbg_processing` set to 1 at all times when still in the debugger
message loop.

The detach finalization bug was triggered by a debug transport callback
called outside the debugger message loop indicating the transport connection
had broken (this would concretely happen e.g. for a periodic Status notify).
This caused detach1() to run which, among other things, sets dbg_read_cb to
NULL and dbg_detaching to 1.

The intent is then for detach2() to finalize the detach (this is separated
from detach1() to allow reattach-during-detached-callback to be handled
correctly).  However, when dbg_read_cb was set to NULL outside the debugger
message loop, the message loop would never be entered again -- and only the
message loop has handling for finalizing a detach using detach2().

The fix for this is to allow the executor to enter the debugger message loop
when a detach is pending (dbg_read_cb == NULL and dbg_detaching == 1) so that
the detach can be correctly finalized.

The message loop detach2() handling now also supports the corner case where
user code reattaches in the detached callback but a detach is immediately
triggered while still inside the detached callback.  This may happen because
`duk_debugger_attach()` writes the debugger protocol handshake line inline
inside the API call using the transport write callback, and that callback
may fail and indicate a transport error.

The fix is to then issue another detached callback etc, until we are either
cleanly detached or cleanly attached and can proceed.

Finally, the control flow was changed so that the `dbg_processing` flag is
set and cleared inside the debugger message loop now rather than in several
separate call sites outside of it.  This should be less error prone.

Both bugs fixed in this commit were introduced in Duktape 1.4.0 as part of
adding the Detaching notify and splitting detach handling into detach1()
and detach2().  As such, the fixes don't need to be backported to Duktape
1.3.x etc.