Use single shift+mask in executor reg/const access
Instead of decoding a field (say A) and then computing DUK__REGP which
involves further shifting of the value, add macros such as DUK__REGP_A(ins)
which operate directly on the 32-bit 'ins' and use a single shift+mask to
get the byte offset (rather than element offset) for the target reg/const.

Technically the compiler should be able to do this, but in practice at
least gcc -O2 won't.