Make plain buffers behave like ArrayBuffers
Change handling of plain buffers so that they behave like ArrayBuffer
instances to Ecmascript code, with limitations such as not being
extensible and all properties being virtualized.  This simplifies
Ecmascript code as plain buffers are just lightweight ArrayBuffers
(similarly to how lightfuncs appear as function objects).  There are
a lot of small changes in how the built-in objects and methods, and
the C API deals with plain buffer values.

Also make a few small changes to plain pointer and lightfunc handling
to improve consistency with how plain buffers are now handled.