Add attr "unused" to internal functs for gcc/clang
When applied to a function declaration the __attribute__ ((unused)
declaration indicates a function may be unused so that the compiler
won't warn about it if it actually is.  This reduces harmless warnings
when a function is unused due to a specific config option combination.

Apply to all Clang versions and GCC >= 3.1.1; GCC 3.1.1 manual confirms
__attribute__ ((unused)) is supported.