Add wrap asserts for INCREF
The refcount field must be chosen so that there's never any wrapping,
otherwise memory unsafe behavior will happen.

For the default configuration a size_t is used for refcount; it's
guaranteed not to wrap because there can't be enough references to
wrap it.  For low memory targets it's possible to configure 16-bit
refcounts.  This is fine as long as memory pools are small enough
to never contain more than 2^16-1 references.