Avoid value stack when throwing a double error
Avoid pushing anything on the value stack when handling a double error.  This
avoids the case where there's no pre-allocated space on the value stack for
pushing the double error.  This is rather theoretical however because there
should always be spare for the error -- but in error paths it's best to have
minimal assumptions.

Also reorganize the error throwing a bit for clarity.