Rework duk_activations into a linked list
Remove thr->callstack as a monolithic array and replace it with a linked list
of duk_activations.  thr->callstack_curr is the current call (or NULL if no
call is in progress), and act->parent chains to a previous call or NULL.
thr->callstack_top is kept because it's needed by some internals at present;
it may be removed in the future.