Avoid duk_hbufobj validity assert trigger
duk_hbufobj validity assertions require that if the buf pointer is NULL,
the buffer length and offset are zero.  This is violated by the shared
slice code because it sets ->length before allocating a buffer, and since an
allocation may trigger side effects, the assert may trigger in mark-and-sweep.
I don't think this causes issues besides an assert failure.