Fix armhf -Wcast-align warnings from Clang and GCC.
These warnings stem from the fact that the double type has 8 byte
alignment on armhf.

The following is from Clang (one of a number of similar warnings):

    content/handlers/javascript/duktape/duktape.c:13390:28: error:
    cast from 'duk_hbuffer *' (aka 'struct duk_hbuffer *')
    to 'duk_hbuffer_fixed *' (aka 'struct duk_hbuffer_fixed *')
    increases required alignment from 4 to 8 [-Werror,-Wcast-align]

This happens because `struct duk_hbuffer_fixed` contains a double in
the union.

The following has a similar cause (again, there are many of these
warnings):

    content/handlers/javascript/duktape/duktape.c:14588:25: error:
    cast from 'duk_hobject *' (aka 'struct duk_hobject *')
    to 'duk_hboundfunc *' (aka 'struct duk_hboundfunc *')
    increases required alignment from 4 to 8 [-Werror,-Wcast-align]

It happens because `struct duk_hboundfunc` contains a `duk_tval`,
and `duk_tval` whihc is a `struct duk_tval_struct`, which contains
a double in the union.

Since all these structures have a `struct duk_heaphdr` at the
start, giving that structure 8-byte alignment fixes the above
warnings.

However, it introduces warnings of the following form:

    content/handlers/javascript/duktape/duktape.c:48342:33: error:
    cast from 'duk_hstring *' (aka 'struct duk_hstring *')
    to 'duk_heaphdr *' (aka 'struct duk_heaphdr *')
    increases required alignment from 4 to 8 [-Werror,-Wcast-align]

Giving `struct duk_hstring` the same 8-byte alignment fixes
all the warnings when building on armhf with both GCC and Clang
with -Wcast-align.