Grok DUK_CONSOLE_TO_STDOUT DUK_CONSOLE_TO_STDERR
Add these two constants, which can be passed as part of the flags to
duk_console_init().  Their meanings are:

- DUK_CONSOLE_TO_STDOUT: all console output goes to stdout
- DUK_CONSOLE_TO_STDERR: all console output goes to stderr
- default, no flags: output for warn and above goes to stderr, output
  for info and below goes to stdout

The constants are used both for the flags parameter to
duk_console_init(), as well as for the flags saved in magic and passed
to duk__console_log_helper(); that way we avoid defining additional,
similar constants for these two different usages.