Fix GETPROPC error handling corner case
When calling via a property and an error happens during call setup, the
error is delayed to be thrown later on, so that proper ECMAScript side
effect ordering semantics are observed.

Previous approach just created the Error object and marked it with a
special _Target symbol marker property.  The code assumed that the Error
was non-callable which would trigger an error path in call handling, and
the special _Target symbol marker would only need to be checked in the
error path.

This overlooked the possibility that a Duktape.errCreate would replace
the non-callable Error object with a callable one, e.g. when one simply
says: Duktape.errCreate = function () { return Float64Array; }.

Modify the approach so that instead of marking the Error object we use
a bare wrapper object which points to the error.  This ensures that
any modifications or replacements to the Error itself won't affect the
call error logic.

Also eliminate one unnecessary call argument from the propcall error
setup helper.