Remove out-of-date checksetup Makefile target
Remove the old 'make checksetup' target and the related script.  The
script is a bit out-of-date, and awkward to maintain for multiple Linux
targets.

Dependencies for util/dist.py are now simple (Python 2 + Python
YAML binding) and explained in the README.  There are also Docker images
for easy development setup which mostly removes the need to document
all the dependencies on all potential Linux distributions for Makefile
targets like building the site.