duk_api_stack: increase memory safety
The ptr_diff is taken between the post reallocation new_valstack pointer
and pre-allocation thr->valstack pointer variables. This is in principle
wrong as the realloc function might change the base pointer, which appears
to be the assumption here that it will not.

Further on this difference is then added to the old base pointers for
bottom/top/end thr->valstack variables, with only thr->valstack being
repalced with the new base pointer. So we have the difference between
two base pointers added to the old base pointer effectively as if the
memory was contiguous.

Lets change that and add the pointer differences to the new_valstack
base pointer obtained from the realloc function instead.

Signed-off-by: Pawel Zalewski <pzalewski@thegoodpenguin.co.uk>